made it basically endless until you press q. big number had to be used because of inability to get it to be infinite

+++ SoundTest.py
@@ -1,6 +1,7 @@
 import pyaudio
 import numpy as np
 import Truncate as tr
+import keyboard
 
 CHUNK = 2**11
 RATE = 44100
@@ -9,14 +10,15 @@ p=pyaudio.PyAudio()
 stream=p.open(format=pyaudio.paInt16,channels=1,rate=RATE,input=True,
               frames_per_buffer=CHUNK)
 #input number for how long you want it to run. Any sounds above the peak(dB) have the time that they were recorded.
-for i in range(int(20.476 * int(input()))): #20.476 * each second
+#for i in range(int(20.476 * int(input()))): #20.476 * each second
+for i in range(100000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000): 
     data = np.fromstring(stream.read(CHUNK),dtype=np.int16)
     peak=np.average(np.abs(data))*2
-    bars="#"*int(50*peak/2**16)
+    bars="#"*int(50*peak/2**16)  
     if (peak/325 > 50):
         print("%s %s "%(tr.truncate((i/20.476),2), bars))
        # print("%04d %05d %s"%((i/20.476)+1,peak/325,bars))
-
-stream.stop_stream()
-stream.close()
-p.terminate()
+    if (keyboard.is_pressed('q')):
+        stream.stop_stream()
+        stream.close()
+        p.terminate()
